tests: Adapt tests to the new client

+++ algoliasearch/types_key.go
@@ -2,13 +2,13 @@ package algoliasearch
 
 type Key struct {
 	ACL                    []string `json:"acl"`
-	CreatedAt              int64    `json:"createdAt,omitempty"`
+	CreatedAt              int      `json:"createdAt,omitempty"`
 	Description            string   `json:"description,omitempty"`
-	MaxHitsPerQuery        int64    `json:"maxHitsPerQuery,omitempty"`
-	MaxQueriesPerIPPerHour int64    `json:"maxQueriesPerIPPerHour,omitempty"`
+	MaxHitsPerQuery        int      `json:"maxHitsPerQuery,omitempty"`
+	MaxQueriesPerIPPerHour int      `json:"maxQueriesPerIPPerHour,omitempty"`
 	QueryParamaters        string   `json:"queryParameters,omitempty"`
 	Referers               []string `json:"referers,omitempty"`
-	Validity               int64    `json:"validity,omitempty"`
+	Validity               int      `json:"validity,omitempty"`
 	Value                  string   `json:"value,omitempty"`
 }
